Distinguish between publication dates and meetup dates
The date property is used as publication date in the RSS feed so as new
meetups are added the publication dates are in the future.

Adding a separate property allows defining independent publication and
meetup dates.

Note: the publication date could be rendered outside of the RSS feed but
currently isn't.